pyproject.toml
# PEP 517 build backend configuration.
[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
    "wheel",
]
build-backend = "setuptools.build_meta"

# PEP 621 core project metadata and runtime requirements.
[project]
name = "polish-pit-calculator"
dynamic = ["version"]
description = "CLI for aggregating brokerage, interest, and manual inputs into Polish PIT summaries."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Adam Rajfer", email = "adam.rajfer@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "prompt-toolkit",
    "pyyaml",
    "questionary",
    "tabulate",
]
keywords = [
    "tax",
    "pit",
    "poland",
    "finance",
    "cli",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
]
scripts = { "pit-pl" = "polish_pit_calculator.app:main" }
urls = { Repository = "https://github.com/AdamRajfer/polish-pit-calculator", Issues = "https://github.com/AdamRajfer/polish-pit-calculator/issues", Changelog = "https://github.com/AdamRajfer/polish-pit-calculator/blob/main/CHANGELOG.md" }

# Optional dependency groups used during development only.
[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "types-PyYAML",
    "types-tabulate",
]

# -----------------------------------------------------------------------------
# Tooling configuration
# -----------------------------------------------------------------------------

# Package discovery for setuptools build backend.
[tool.setuptools.packages.find]
include = ["polish_pit_calculator", "polish_pit_calculator.*"]

# Derive package version from git tags.
[tool.setuptools_scm]
fallback_version = "0.0.0"

# uv project behavior.
[tool.uv]
package = true

# Code cleanup tooling (remove unused imports/variables).
[tool.autoflake]
remove_all_unused_imports = true
in_place = true
remove_unused_variables = true

# Code formatter configuration.
[tool.black]
line-length = 100

# Import sorting configuration.
[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

# Pylint: keep strict checks, but allow data-heavy model classes.
[tool.pylint.design]
max-attributes = 16
max-args = 6
max-positional-arguments = 6
max-locals = 25
min-public-methods = 1

[tool.pylint.format]
max-module-lines = 1400

# Pytest defaults, including coverage reporting.
[tool.pytest.ini_options]
addopts = "--strict-config --strict-markers --cov=polish_pit_calculator --cov-branch --cov-report=term-missing --cov-fail-under=100"

# Coverage.py behavior used by pytest-cov.
[tool.coverage.run]
branch = true
source = ["polish_pit_calculator"]

[tool.coverage.report]
show_missing = true
fail_under = 100
precision = 2
exclude_lines = [
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
